Update to common functions to include stopwatch class
